requirements.txt
langchain>=0.1.0
langchain-anthropic>=0.1.0
python-binance>=1.0.34
pandas>=2.0.0
numpy>=1.24.0
python-dotenv>=1.0.0
loguru>=0.7.0
requests>=2.31.0
